Tests/LibWeb/Text/input/HTML/document-close-iframe-load-event.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(done => {
        const iframe = document.createElement('iframe');
        document.body.appendChild(iframe);

        const iframeDocument = iframe.contentDocument;
        iframeDocument.open().write('<body onload="window.onBodyLoaded(document)"></body>');
        iframeDocument.addEventListener("readystatechange", () => {
            println(`Ready state changed to ${iframeDocument.readyState}`);
        });
        iframeDocument.addEventListener("DOMContentLoaded", () => {
            println("DOMContentLoaded event fired");
        });
        iframe.contentWindow.onBodyLoaded = document => {
            println(`Body onload event fired with ready state ${document.readyState}`);
        };
        iframe.onload = () => {
            println("Iframe onload event fired");
            iframe.remove();
            done();
        };
        iframeDocument.close();
        iframeDocument.close();
    });
</script>
