Tests/LibWeb/Text/expected/HTML/document-close-iframe-load-event.txt

Ready state changed to interactive
DOMContentLoaded event fired
Ready state changed to complete
Body onload event fired with ready state complete
Iframe onload event fired
